backend/requirements.txt
cohere==4.2.0
qdrant-client==1.9.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
pytest==7.4.0
openai>=1.0.0
openai-agents>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0